+++ src/compiler/regalloc.h
@@ -1,9 +1,0 @@
-#ifndef HAVE_REGALLOC_H
-#define HAVE_REGALLOC_H
-
-#include "codegen.h"
-
-bool RA_regAlloc (CG_frame f, AS_instrList il);
-
-#endif
-